+++ integration_tests/test_notification2.py
@@ -1,18 +1,55 @@
 # Copyright (c) 2025 Cumulocity GmbH
 
+# pylint: disable=protected-access
+
 import asyncio
+import contextlib
+import logging
 import queue
 import threading
 import time
+from unittest.mock import Mock
 
 import pytest
 
 from c8y_api import CumulocityApi
-from c8y_api.model import Device, ManagedObject, Subscription, Measurement, Value, Event, Alarm, Operation
+from c8y_api.model import Device, ManagedObject, Subscription, Measurement, Value, Event, Alarm
 from c8y_tk.notification2 import AsyncListener, Listener
-from tests.utils import assert_in_any
+from tests.utils import assert_in_any, assert_no_failures, assert_all_in
 
 from util.testing_util import RandomNameGenerator
+
+# @pytest.fixture(autouse=True)
+# def no_stray_threads():
+#     before = {t.name for t in threading.enumerate() if t.is_alive()}
+#     yield
+#     # small grace for shutdowns to finish
+#     time.sleep(0.05)
+#     after = {t.name for t in threading.enumerate() if t.is_alive()}
+#
+#     # ignore main & pytest internals
+#     ignored = {t.name for t in threading.enumerate() if t.daemon and t.name.startswith(("pytest", "MainThread"))}
+#     leaked = [n for n in after - before if n not in ignored]
+#
+#     assert not leaked, f"Leaked threads: {leaked}"
+#
+#
+
+@pytest.fixture(scope="function", name="sample_subscription")
+def fix_sample_subscription(sample_object):
+    """Provide a "catch-all" subscription for a sample object."""
+    sub = Subscription(
+        c8y=sample_object.c8y,
+        name=f'{sample_object.name.replace("_", "")}Subscription',
+        context=Subscription.Context.MANAGED_OBJECT,
+        api_filter=['*'],
+        source_id=sample_object.id,
+    ).create()
+
+    yield sub
+
+    with contextlib.suppress(Exception):
+        sub.delete()
 
 
 def test_subscription_deletion(live_c8y, safe_create):
@@ -25,6 +62,7 @@ def test_subscription_deletion(live_c8y, safe_create):
 
     assert live_c8y.notification2_subscriptions.get_count(source=mo.id) == 1
     mo.delete()
+    time.sleep(1)
     assert live_c8y.notification2_subscriptions.get_count(source=mo.id) == 0
 
 
@@ -48,6 +86,7 @@ def fix_object_tree_builder(live_c8y: CumulocityApi, safe_create):
     return build
 
 
+# TODO: Add Operation
 @pytest.mark.parametrize("api_filters, expected", [
     ('*', 'M,E,EwC,A,AwC,MO'),
     ('M', 'M'),
@@ -72,7 +111,6 @@ def test_api_filters(live_c8y: CumulocityApi, sample_object, api_filters, expect
     a couple of corresponding changes. It then matches the received
     notifications against expectations.
     """
-    # TODO: Add Operation
     apis = {
         '*': '*',
         'M': 'measurements',
@@ -110,10 +148,12 @@ def test_api_filters(live_c8y: CumulocityApi, sample_object, api_filters, expect
         time.sleep(3)  # ensure creation
 
         # (2) apply updates
-        m_id = Measurement(live_c8y, source=mo.id, type="c8y_TestMeasurement", metric=Value(1, '')).create()
-        e_id = Event(live_c8y, source=mo.id, type="c8y_TestEvent", time='now', text='text').create()
-        a_id = Alarm(live_c8y, source=mo.id, type="c8y_TestAlarm", time='now', text='text', severity=Alarm.Severity.WARNING).create()
-        # o_id = Operation(live_c8y, device_id=mo.id, c8y_Operation={}).create()
+        m_id = Measurement(live_c8y, source=mo.id, type="c8y_TestMeasurement",
+                           metric=Value(1, '')).create().id
+        e_id = Event(live_c8y, source=mo.id, type="c8y_TestEvent", time='now', text='text').create().id
+        a_id = Alarm(live_c8y, source=mo.id, type="c8y_TestAlarm", time='now', text='text',
+                     severity=Alarm.Severity.WARNING).create().id
+        # o_id = Operation(live_c8y, device_id=mo.id, c8y_Operation={}).create().id
         mo.apply({'some_tag': {}})
 
         time.sleep(1)
@@ -121,12 +161,110 @@ def test_api_filters(live_c8y: CumulocityApi, sample_object, api_filters, expect
         # collect message types from source URL
         types = {n.source.split('/')[2] for n in ns}
         for e in expected:
-            assert_in_any(e, *types)
-
+            assert_in_any(e, types)
+        ids = {n.json['id'] for n in ns}
+        if 'measurements' in expected:
+            assert m_id in ids
+        if 'events' in expected:
+            assert e_id in ids
+        if 'alarms' in expected:
+            assert a_id in ids
     finally:
         # (99) cleanup
-        listener.close()
+        listener.stop()
         listener_thread.join()
+
+
+@pytest.mark.asyncio(loop_scope='function')
+async def test_asyncio_token_timeout(caplog, live_c8y: CumulocityApi, sample_subscription):
+    """Verify that token timeout is handled properly.
+
+    This test configures the listener to use a one-minute token timeout. A
+    single change is applied to the monitored managed object to verify that
+    the setup was successful. The test then waits for the token to time out
+    before stopping the listener.
+
+    -> notification was handled
+    -> two tokens were generated
+    """
+    caplog.set_level(logging.INFO)
+
+    obj_id = sample_subscription.source_id
+    callback = Mock()
+
+    async def async_callback(msg):
+        callback()
+
+    listener = AsyncListener(
+        live_c8y,
+        subscription_name=sample_subscription.name,
+        auto_ack=True,
+    )
+    listener.token_validity = 1  # minimize validity
+    listener.start(async_callback)
+    await asyncio.sleep(5)  # ensure creation
+    token1 = listener._current_token
+    try:
+        live_c8y.inventory.apply_to({'test_CustomFragment': {'num': 42}}, obj_id)
+        await asyncio.sleep(20)  # wait for the token to expire
+    finally:
+        listener.stop()
+        await listener.wait()
+    token2 = listener._current_token
+
+    # -> callback was invoked once
+    callback.assert_called_once()
+    # -> 2 tokens where generated
+    assert token1 != token2
+    # -> 2 corresponding log messages
+    log_messages = [r.message for r in caplog.records]
+    assert sum("Notification 2.0 token requested" in x for x in log_messages) == 2
+    assert sum("cancelled" in x for x in log_messages) == 1
+    assert_no_failures(caplog)
+
+
+def test_token_timeout(caplog, live_c8y: CumulocityApi, sample_subscription):
+    """Verify that token timeout is handled properly.
+
+    This test configures the listener to use a one-minute token timeout. A
+    single change is applied to the monitored managed object to verify that
+    the setup was successful. The test then waits for the token to time out
+    before stopping the listener.
+
+    -> notification was handled
+    -> two tokens were generated
+    """
+    caplog.set_level(logging.INFO)
+
+    obj_id = sample_subscription.source_id
+    callback = Mock()
+
+    listener = Listener(
+        live_c8y,
+        subscription_name=sample_subscription.name,
+        auto_ack=True,
+    )
+    listener.token_validity = 1  # minimize validity
+    listener.start(callback)
+    time.sleep(5)  # ensure creation
+    token1 = listener._listener._current_token
+    try:
+        live_c8y.inventory.apply_to({'test_CustomFragment': {'num': 42}}, obj_id)
+        time.sleep(20)  # wait for the token to expire
+    finally:
+        listener.stop()
+        listener.wait()
+    token2 = listener._listener._current_token
+
+    # -> callback was invoked once
+    callback.assert_called_once()
+    # -> 2 tokens where generated
+    assert token1 != token2
+    # -> 2 corresponding log messages
+    log_messages = [r.message for r in caplog.records]
+    assert sum("Notification 2.0 token requested" in x for x in log_messages) == 2
+    assert sum("cancelled" in x for x in log_messages) == 1
+    assert_no_failures(caplog)
 
 
 def test_object_update_and_deletion(live_c8y: CumulocityApi, sample_object):
@@ -193,7 +331,7 @@ def test_object_update_and_deletion(live_c8y: CumulocityApi, sample_object):
     assert m.action == "DELETE"
 
     # (99) cleanup
-    listener.close()
+    listener.stop()
     listener_thread.join()
 
 
@@ -221,13 +359,13 @@ async def test_asyncio_object_update_and_deletion(logger, live_c8y: CumulocityAp
         await m.ack()
 
     listener = AsyncListener(live_c8y, subscription_name=sub.name)
-    listener_task = asyncio.create_task(listener.listen(receive_notification))
+    listener.start(receive_notification)
     await asyncio.sleep(5)  # ensure creation
 
     # (1) apply first change, expected fragment
     mo.apply({'test_AwaitedFragment': {'num': 42}})
     # -> notification should appear
-    m = await notifications.get()
+    m = await asyncio.wait_for(notifications.get(), timeout=1)
     assert notifications.empty()
     assert mo.id in m.source
     assert m.action == "UPDATE"
@@ -240,7 +378,7 @@ async def test_asyncio_object_update_and_deletion(logger, live_c8y: CumulocityAp
     # (2) Apply 2nd change, different fragment
     mo.apply({'test_DifferentFragment': {'num': 42}})
     # -> notification should appear
-    m = await notifications.get()
+    m = await asyncio.wait_for(notifications.get(), timeout=1)
     assert notifications.empty()
     assert mo.id in m.source
     assert m.action == "UPDATE"
@@ -255,14 +393,14 @@ async def test_asyncio_object_update_and_deletion(logger, live_c8y: CumulocityAp
     # (3) delete object tree
     mo.delete_tree()
     # -> notification should appear
-    m = await notifications.get()
+    m = await asyncio.wait_for(notifications.get(), timeout=1)
     assert notifications.empty()
     assert mo.id in m.source
     assert m.action == "DELETE"
 
     # (99) cleanup
-    await listener.close()
-    await listener_task
+    listener.stop()
+    await listener.wait()
 
 
 def test_child_updates(live_c8y: CumulocityApi, object_tree_builder):
@@ -282,23 +420,22 @@ def test_child_updates(live_c8y: CumulocityApi, object_tree_builder):
         m.ack()
 
     listener = Listener(live_c8y, subscription_name=root_subscription.name)
-    listener_thread = threading.Thread(target=listener.listen, args=[receive_notification])
-    listener_thread.start()
-    time.sleep(5)  # ensure creation
-
-    # update all child objects
-    live_c8y.inventory.apply_to(
-        {'test_CustomFragment': {'num': 42}},
-        *[x.id for x in root.child_assets],
-        *[x.id for x in root.child_devices],
-        *[x.id for x in root.child_additions],
-    )
+    try:
+        listener.start(receive_notification)
+        time.sleep(5)  # ensure creation
+        # update all child objects
+        live_c8y.inventory.apply_to(
+            {'test_CustomFragment': {'num': 42}},
+            *[x.id for x in root.child_assets],
+            *[x.id for x in root.child_devices],
+            *[x.id for x in root.child_additions],
+        )
+    finally:
+        # (99) cleanup
+        listener.stop()
+        listener.wait()
 
     assert not notified.is_set()
-
-    # (99) cleanup
-    listener.close()
-    listener_thread.join()
 
 
 @pytest.mark.asyncio(loop_scope='function')
@@ -333,7 +470,7 @@ async def test_asyncio_child_updates(live_c8y: CumulocityApi, object_tree_builde
     assert not notifications
 
     # (99) cleanup
-    await listener.close()
+    listener.stop()
     await listener_task
 
 
@@ -358,21 +495,19 @@ def test_parent_updates(live_c8y: CumulocityApi, object_tree_builder):
         m.ack()
 
     listeners = [Listener(live_c8y, subscription_name=s.name) for s in child_subscriptions]
-    listener_threads = [threading.Thread(target=l.listen, args=[receive_notification]) for l in listeners]
-    for l in listener_threads:
-        l.start()
-    time.sleep(5)  # ensure creation
-
-    # update all child objects
-    root.apply({'test_CustomFragment': {'num': 42}})
-
+    try:
+        for l in listeners:
+            l.start(receive_notification)
+        time.sleep(5)  # ensure creation
+        # update all child objects
+        root.apply({'test_CustomFragment': {'num': 42}})
+    finally:
+        # (99) cleanup
+        for l in listeners:
+            l.stop()
+        for l in listeners:
+            l.wait()
     assert not notifications
-
-    # (99) cleanup
-    for l in listeners:
-        l.close()
-    for l in listener_threads:
-        l.join()
 
 
 @pytest.mark.asyncio(loop_scope='function')
@@ -406,8 +541,10 @@ async def test_asyncio_parent_updates(live_c8y: CumulocityApi, object_tree_build
     assert not notifications
 
     # (99) cleanup
-    await asyncio.gather(*[l.close() for l in listeners])
-    await asyncio.wait(listener_tasks)
+    for l in listeners:
+        l.stop()
+    for l in listener_tasks:
+        await l
 
 
 def build_managed_object_subscription(mo):
@@ -424,13 +561,15 @@ def create_managed_object_subscription(c8y, mo):
     return s.create()
 
 
-def test_multiple_subscribers(live_c8y: CumulocityApi, sample_object):
+@pytest.mark.parametrize("shared", [True, False])
+def test_multiple_subscribers(caplog, live_c8y: CumulocityApi, sample_object, shared):
     """Verify that multiple subscribers/consumers can be created for a single subscription.
 
     This test creates a managed object and corresponding subscription as well as multiple
     listeners with unique subscriber names. An update to the managed object should notify
     each of the subscribers.
     """
+    caplog.set_level(logging.DEBUG)
 
     mo = sample_object
     sub = create_managed_object_subscription(live_c8y, mo)
@@ -442,58 +581,88 @@ def test_multiple_subscribers(live_c8y: CumulocityApi, sample_object):
         m.ack()
 
     n_listeners = 3
-    listeners = [Listener(live_c8y, subscription_name=sub.name, subscriber_name=f"{sub.name}{i}")
-                 for i in range(n_listeners)]
-    listener_threads = [threading.Thread(target=l.listen, args=[receive_notification]) for l in listeners]
-    for l in listener_threads:
-        l.start()
+    listeners = [
+        Listener(
+            live_c8y,
+            subscription_name=sub.name,
+            subscriber_name=f"{sub.name}{i if not shared else 0}",
+            shared=shared,
+        )
+        for i in range(n_listeners)
+    ]
+    listener_threads = [l.start(receive_notification) for l in listeners]
     time.sleep(5)  # ensure creation
 
-    # update the object
-    mo.apply({'test_CustomFragment': {'num': 42}})
-    time.sleep(5)  # ensure processing
-    # -> all received notifications are identical
-    assert len(notifications) == 3
-    assert len({n.raw for n in notifications}) == 1
-
-    # (99) cleanup
-    for l in listeners:
-        l.close()
-    for l in listener_threads:
-        l.join()
+    try:
+        # update the object
+        mo.apply({'test_CustomFragment': {'num': 42}})
+        time.sleep(5)  # ensure processing
+        # -> there should be 1/3 messages for shared/not shared
+        assert len(notifications) == (3 if not shared else 1)
+        # -> all received notifications are identical
+        assert len({n.raw for n in notifications}) == 1
+    finally:
+        for l in listeners:
+            l.stop()
+        t0 = time.time()
+        for l in listener_threads:
+            l.join()
+        t1 = time.time()
+        assert t1 - t0 < 10
 
 
 @pytest.mark.asyncio(loop_scope='function')
-async def test_asyncio_multiple_subscribers(live_c8y: CumulocityApi, sample_object):
+@pytest.mark.parametrize("shared", [True, False])
+async def test_asyncio_multiple_subscribers(caplog, live_c8y: CumulocityApi, sample_object, shared):
     """Verify that multiple subscribers/consumers can be created for a single subscription.
 
     This test creates a managed object and corresponding subscription as well as multiple
-    listeners with unique subscriber names. An update to the managed object should notify
-    each of the subscribers.
+    listeners with shared/unique subscriber names. An update to the managed object should
+    notify each of the subscribers if unique or just one of the subscribers if shared.
     """
+    caplog.set_level(logging.INFO)
 
     mo = sample_object
     sub = create_managed_object_subscription(live_c8y, mo)
 
-    # prepare listeners
+    # feedback callback
     notifications:list[AsyncListener.Message] = []
     async def receive_notification(m:AsyncListener.Message):
         notifications.append(m)
         await m.ack()
 
+    # prepare listeners
     n_listeners = 3
-    listeners = [AsyncListener(live_c8y, subscription_name=sub.name, subscriber_name=f"{sub.name}{i}")
-                 for i in range(n_listeners)]
-    listener_tasks = [asyncio.create_task(l.listen(receive_notification)) for l in listeners]
-    await asyncio.sleep(5)  # ensure creation
+    listeners = [
+        AsyncListener(
+            live_c8y,
+            subscription_name=sub.name,
+            subscriber_name=f"{sub.name}{i if not shared else 0}",
+            shared=shared,
+        )
+        for i in range(n_listeners)
+    ]
 
-    # update the object
-    mo.apply({'test_CustomFragment': {'num': 42}})
-    await asyncio.sleep(5)  # ensure processing
+    for l in listeners:
+        l.start(receive_notification)
+    try:
+        await asyncio.sleep(5)  # ensure creation
+        mo.apply({'test_CustomFragment': {'num': 42}})
+        await asyncio.sleep(3)  # ensure processing
+    finally:
+        for l in listeners:
+            l.stop()
+        for l in listeners:
+            await l.wait()
+
+    # -> there should be 1/3 messages for shared/not shared
+    assert len(notifications) == (3 if not shared else 1)
     # -> all received notifications are identical
-    assert len(notifications) == 3
     assert len({n.raw for n in notifications}) == 1
-
-    # (99) cleanup
-    await asyncio.gather(*[l.close() for l in listeners])
-    await asyncio.wait(listener_tasks)
+    # -> if shared, there should be unsubscribed infos/warnings
+    log_messages = [r.message for r in caplog.records]
+    assert sum("unsubscribed." in x for x in log_messages) == (1 if shared else n_listeners)
+    assert (sum("could not be unsubscribed (assuming it was already unsubscribed)" in x for x in log_messages)
+            == (n_listeners-1 if shared else 0))
+    assert sum("cancelled" in x for x in log_messages) == n_listeners
+    assert_no_failures(caplog)